+++ build.gradle
@@ -28,7 +28,7 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.duke.Duke")
+    mainClass.set("seedu.duke.duke.main.Duke")
 }
 
 shadowJar {
