+++ src/main/java/duke/command/ByeCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.Storage;
+import duke.main.Storage;
 import duke.task.TaskList;
 
 public class ByeCommand extends Command {
