+++ src/main/java/duke/command/Command.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.Storage;
+import duke.main.Storage;
 import duke.task.TaskList;
 
 public abstract class Command {
