+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.Storage;
+import duke.main.Storage;
 import duke.task.TaskList;
 
 public class UnmarkCommand extends Command {
