+++ src/main/java/duke/exception/DukeException.java
@@ -1,6 +1,6 @@
 package duke.exception;
 
-/** This class handles exceptions from duke.Duke.*/
+/** This class handles exceptions from duke.main.Duke.*/
 public class DukeException extends Exception {
     public DukeException(String errorMsg) {
         super(errorMsg);
