+++ src/main/java/duke/main/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.main;
 
 public class Ui {
     public void greetUser() {
